+++ CHANGELOG.md
@@ -10,9 +10,11 @@
 
 ### 🛠 Improvements
 - [6942](https://github.com/vegaprotocol/vega/issues/6942) - Add `admin.rename_network` with `vega wallet network rename`
+- [7288](https://github.com/vegaprotocol/vega/issues/7288) - Add `block` interval for trade candles
 
 ### 🐛 Fixes
 - [7535](https://github.com/vegaprotocol/vega/issues/7535) - Fix network history load takes an increasingly long time to complete
+- [7517](https://github.com/vegaprotocol/vega/issues/7517) - Add buffer files event source 
 - [7659](https://github.com/vegaprotocol/vega/issues/7659) - Tidy up REST documentation for consistency
 - [](https://github.com/vegaprotocol/vega/issues/xxxx) -